Added usergroup to user component.

+++ components/cookbooks/user/recipes/delete.rb
@@ -1,15 +1,24 @@
-username = node[:user][:username]
-group = JSON.parse(node[:user][:group])
+usergroup = node['user']['usergroup'] == 'true'
+if usergroup
+  users = JSON.parse(node['user']['usermap'])
+else
+  keys_string = JSON.parse(node['user']['authorized_keys']).join("\n")
+  users = { node['user']['username'] => keys_string }
+end
+group = JSON.parse(node['user']['group'])
 is_windows = node['platform'] =~ /windows/
 group.push('Administrators') if is_windows && !group.include?('Administrators')
 
 if !is_windows
   Chef::Log.info("Stopping the nslcd service")
-  `sudo killall -9  /usr/sbin/nslcd`
+  `sudo pkill -9  /usr/sbin/nslcd`
+end
 
-  if username != "root"
-    execute "pkill -9 -u #{username} ; true"
-  end
+
+users.each do |username, keys|
+
+if username != "root" && !is_windows
+  execute "pkill -9 -u #{username} ; true"
 end
 
 user username do
@@ -29,3 +38,5 @@ group.each do |g|
     action :manage
   end
 end if is_windows
+
+end
